requirement.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
uuid==1.30
